Refactor Internals to Tonal Interface
