+++ app/src/components/Seizure.jsx
@@ -1,32 +1,11 @@
 import React from 'react';
-import { connect } from 'react-redux';
-import mapDispatchToProps from '../redux/dispatch/seizure';
-import mapStateToProps from '../redux/selectors/seizure';
 import Panel from './Panel';
+import TonalInterface from '../interfaces/tonal';
 
-const audioContext = new AudioContext();
 
 class Seizure extends React.Component {
-  constructor(props) {
-    super(props);
-    props.init();
-  }
-
-  _generateTone(n) {
-    let oscillator = audioContext.createOscillator();
-
-    oscillator.connect(audioContext.destination);
-
-    oscillator.type = 'sine';
-    oscillator.frequency.value = `${n}${n}0`;
-    oscillator.detune.value = `${n}00`;
-
-    oscillator.start(audioContext.currentTime);
-    oscillator.stop(audioContext.currentTime + .25);
-  }
-
   render() {
-    const { colors = [], randomizeColors } = this.props;
+    const { colors = [], randomizeColors, generateTone } = this.props;
     return (
       <div className="flex-container-column">
         <button onClick={randomizeColors}>
@@ -36,7 +15,7 @@ class Seizure extends React.Component {
           {colors.map((color, n) => (
             <Panel
               className={`${color} panel`}
-              onMouseOver={() => this._generateTone(n)}
+              onMouseOver={() => generateTone(n)}
               />
           ))}
         </div>
@@ -44,8 +23,4 @@ class Seizure extends React.Component {
     );
   }
 }
-
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps,
-)(Seizure)
+  export default TonalInterface(Seizure);
