+++ app/src/utils/getAudioContext.js
@@ -1,0 +1,8 @@
+let audioContext = null;
+
+export default function getAudioContext() {
+  if (!audioContext) {
+    audioContext = new AudioContext();
+  }
+  return audioContext;
+}
